Update .gitattributes
Since Scoop is a Windows-only tool, we can safely use CRLF line endings for all text files.
If Git decides that the content is text, its line endings will be normalized to CRLF in the working tree on checkout.
In the Git index/repository the files will always be stored with LF line endings. This is fine.